feat: 좋아요 수 안 뜨는 오류 수정
